Create a function to initialize application in __init__.py in application package
