+++ application/views/__init__.py
@@ -1,10 +1,2 @@
-from application import jwt
 
-# To be used for storing blacklisted tokens
 blacklist = set()
-
-
-@jwt.token_in_blacklist_loader
-def check_if_token_in_blacklist(decrypted_token):
-    jti = decrypted_token['jti']
-    return jti in blacklist
