+++ tests/testsbookretrieval.py
@@ -1,8 +1,8 @@
 import json
 import unittest
 
-from run import *
 from application.models.bookmodels import *
+from application import *
 
 
 class TestsBook(unittest.TestCase):
@@ -12,8 +12,7 @@ class TestsBook(unittest.TestCase):
         Register a user; an admin and use the admin details to add a new book to database.
 
         """
-        self.app = app
-        self.app.config.from_object(configuration['testing'])
+        self.app = app_config('testing')
         self.app = self.app.test_client()
         db.create_all()
         url = '/api/v1/auth/'
