+++ tests/testsborrowbook.py
@@ -1,7 +1,7 @@
 import json
 import unittest
 
-from run import *
+from application import *
 
 
 class TestsBook(unittest.TestCase):
@@ -11,8 +11,7 @@ class TestsBook(unittest.TestCase):
         Register a user; an admin and use the admin details to add a new book to database.
 
         """
-        self.app = app
-        self.app.config.from_object(configuration['testing'])
+        self.app = app_config('testing')
         self.app = self.app.test_client()
         db.create_all()
         url = '/api/v1/auth/'
